set('online') should not subscribe
- By removing this implicit subscription,
we can now have presence resources where the 'present'
clients are not interested in the overall presences.
This allows 1000s of clients to go online on the same presence,
and makes it possible to read their collective state through API,
or allow limited number of subscribers to such resources (like
dashboards).